+++ tests/Fakes/AbstractData/AbstractData.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Spatie\LaravelData\Tests\Fakes\AbstractData;
+
+use Spatie\LaravelData\Data;
+
+abstract class AbstractData extends Data
+{
+}