+++ tests/Support/EloquentCasts/DataEloquentCastTest.php
@@ -4,10 +4,11 @@ use Illuminate\Support\Facades\DB;
 
 use function Pest\Laravel\assertDatabaseHas;
 
+use Spatie\LaravelData\Support\DataConfig;
+use Spatie\LaravelData\Tests\Fakes\AbstractData\AbstractDataA;
+use Spatie\LaravelData\Tests\Fakes\AbstractData\AbstractDataB;
 use Spatie\LaravelData\Tests\Fakes\Models\DummyModelWithCasts;
-
 use Spatie\LaravelData\Tests\Fakes\Models\DummyModelWithDefaultCasts;
-
 use Spatie\LaravelData\Tests\Fakes\SimpleData;
 use Spatie\LaravelData\Tests\Fakes\SimpleDataWithDefaultValue;
 
@@ -78,4 +79,54 @@ it('loads a cast object when nullable argument used and value is null in databas
     expect($model->data)
         ->toBeInstanceOf(SimpleDataWithDefaultValue::class)
         ->string->toEqual('default');
+});
+
+it('can use an abstract data class with multiple children', function () {
+    $abstractA = new AbstractDataA('A\A');
+    $abstractB = new AbstractDataB('B\B');
+
+    $modelId = DummyModelWithCasts::create([
+        'abstract_data' => $abstractA,
+    ])->id;
+
+    $model = DummyModelWithCasts::find($modelId);
+
+    expect($model->abstract_data)
+        ->toBeInstanceOf(AbstractDataA::class)
+        ->a->toBe('A\A');
+
+    $model->abstract_data = $abstractB;
+    $model->save();
+
+    $model = DummyModelWithCasts::find($modelId);
+
+    expect($model->abstract_data)
+        ->toBeInstanceOf(AbstractDataB::class)
+        ->b->toBe('B\B');
+});
+
+it('can use an abstract data class with morph map', function () {
+    app(DataConfig::class)->enforceMorphMap([
+        'a' => AbstractDataA::class,
+    ]);
+
+    $abstractA = new AbstractDataA('A\A');
+    $abstractB = new AbstractDataB('B\B');
+
+    $modelA = DummyModelWithCasts::create([
+        'abstract_data' => $abstractA,
+    ]);
+
+    $modelB = DummyModelWithCasts::create([
+        'abstract_data' => $abstractB,
+    ]);
+
+    expect(json_decode($modelA->getRawOriginal('abstract_data'))->type)->toBe('a');
+    expect(json_decode($modelB->getRawOriginal('abstract_data'))->type)->toBe(AbstractDataB::class);
+
+    $loadedMorphedModel = DummyModelWithCasts::find($modelA->id);
+
+    expect($loadedMorphedModel->abstract_data)
+        ->toBeInstanceOf(AbstractDataA::class)
+        ->a->toBe('A\A');
 });